extract RewrittenPackages
